[v0.2.0] WP2 — Power, ADC, temperature, EEPROM, battery
Adds the full main-board measurement and power-management chain. No sensor
daughterboards yet — they remain stubbed.

Implemented:
- hal_adc: 4-channel DMA scan (BATTERY_SENSE, VREFINT, TEMP_SENSE_EXT,
  TEMP_SENSE) into a static AdcResults struct; HAL_ADC_ConvCpltCallback
  drives the ready flag
- hal_i2c: HAL_I2C_MAIN with blocking + DMA primitives, write-cycle
  polling via HAL_I2C_IsDeviceReady; HAL_I2C_AUX remains stubbed
- drv_lm35: ADC raw → 0.01°C using LM35_SCALE = 5 (10 mV/°C, 2.048 V
  VREFBUF, 12-bit) — formula not yet re-verified against the REV B
  TMP236 on-board sensor, see pin_config.h
- drv_24lc256: page-aligned write, 2-byte address-then-DMA reads,
  internal state machine with t_WC ACK polling driven by
  drv_24lc256_update()
- math_crc16: CRC16-CCITT (poly 0x1021, init 0xFFFF), table-driven,
  table built lazily on first use
- svc_storage: load → magic + version + CRC verify → migrate or
  reseed defaults; non-blocking multi-page writes that respect the
  64-byte page boundary; CalibrationData independent of DeviceSettings
- svc_battery: SOC LUT (11-point LiPo OCV curve), state classification
  (NORMAL/LOW/CRITICAL/CHARGING/FULL), 2-second deferred shutdown
  latch that drops the 3.3 V rail
- app_scheduler: full WP2 task table (adc, temperature, battery,
  storage, display) with periods sourced from g_device_settings
- app_display: status screen with temperature, battery %, voltage,
  CHARGING/USB badges; dedicated low-battery shutdown screen

Boot sequence in main.c USER CODE BEGIN 1 asserts the 3.3 V rail enable
before HAL_Init using LL inline functions, then a ~5 ms calibration loop.
USER CODE BEGIN 2 calls hal_*/drv_*/svc_* init in dependency order, with
svc_storage_init() ahead of app_scheduler_init() so task periods come
from loaded settings.

CalibrationData added to system_state with global g_calibration; LM35
and battery monitoring constants live in pin_config.h.

Rebased onto the REV B pinout (was REV A / pre-migration when originally
written); the CubeMX-regen-only commit from the original branch was
dropped since REV B's CubeMX config already covers WP2-5. Fixes beyond
mechanical pin renames, made while rebasing:
- 3V3_EN (was LDO_EN) inverted polarity: REV B's !3V3_EN! is active-LOW
  (old LDO_EN was active-HIGH) — both the pre-HAL_Init boot code in
  main.c and svc_battery.c's shutdown latch now assert/clear the correct
  level, not just the renamed pin.
- ADC1 scan order: REV B's fixed/NOT_FULLY_CONFIGURABLE sequencer always
  converts channels in ascending channel-number order (CH3, CH13/VREFINT,
  CH15, CH16), not the GUI's rank order — hal_adc.c's DMA buffer indices
  now match the real hardware order, not the old 3-channel REV A layout.
- BATTERY_SENSE/TEMP_SENSE moved to PA3/PB12 (ADC1_IN3/IN16, was PB0/PB1
  IN8/IN9); CHARGE_SENSE moved to PC2 (was PA9); VBUS_SENSE moved to PA2
  (was PA10); LED_PWR/LED_STATUS moved to PB13/PB14 (was PC1/PC2).
- VBUS_SENSE (PA2) is a SYS_WKUP4 pin with no CubeMX-generated GPIO_Init
  call, left in POR-default Analog mode — hal_gpio_init() now explicitly
  reconfigures it as a digital input so svc_battery can actually read it.
- 5V_EN (PC7, new on REV B) is asserted unconditionally at boot alongside
  3V3, since TEMP_SENSE/TEMP_SENSE_EXT need it enabled to read correctly
  and no reference-counted power sequencing exists yet.
- hal_spi.c (WP1, not WP2, but touched here): display moved from SPI1 to
  SPI2 in REV B (SPI1 is now the external ADC front-end) — was still
  wired to hspi1/SPI1, silently talking to the wrong peripheral.

Known gap, not fixed here: hal_tim.c's hal_tim_vcom_start() still drives
TIM3_CH1 hardware PWM for the display VCOM signal. REV B removed VCOM's
hardware timer channel entirely (PD3 is plain GPIO_Output; TIM3_CH4 is
committed to the buzzer) — this compiles but does not actually toggle
VCOM, which per CLAUDE.md permanently damages the display if it runs
unmodified. Needs a TIM6-ISR-driven manual toggle (already speced in
docs/cubemx_configuration_checklist.md §2) before any hardware bring-up;
out of scope for this rebase since it's WP1-owned, not part of any
wp2-5 diff.

+++ Math/math_crc.h
@@ -3,6 +3,7 @@
 
 #include <stdint.h>
 
-uint16_t math_crc16(const uint8_t *data, uint32_t len);
+/* CRC16-CCITT (poly 0x1021, init 0xFFFF, no reflection, no XOR-out) */
+uint16_t math_crc16(const uint8_t *data, uint16_t len);
 
 #endif /* MATH_CRC_H */
